requirements.txt
streamlit
crewai
langchain
langchain-openai
python-dotenv
openai
wikipedia
duckduckgo-search
